refactor(dsp): remove C++ analyzer suppressions (fix underlying code)
Codebase-wide removal of clang-tidy/cppcheck suppressions, re-evaluating each
from a fresh lens (SME + the-fool) and fixing the underlying code rather than
re-suppressing. Only one suppression remains, and it is genuinely irreducible.

Logging seam (drops all cppcoreguidelines-pro-type-vararg suppressions):
- Add AsLog.h — off-RT std::format + std::fwrite seam. The format string is
  consteval-validated and it is a variadic template (not C varargs), so it is
  vararg-suppression-free and safe under -fno-exceptions.
- Route every fprintf/NSLog/snprintf through it: Realizer, AUAudioUnit,
  CoreAudioDevice, FileDecodeSource, HALOutputEngine (FourCC -> format_to_n).

Enum-size (drops performance-enum-size suppressions):
- AUParameterID uint64_t -> uint8_t; CrossfeedPreset dead-NOLINT removed.

Analyzer fixes:
- FileDecodeSource: the clang-analyzer-core.CallAndMessage suppression was
  stale (the current analyzer does not fire); deleted + documented the
  loaded-invariant instead.
- LoudnessModule: reorder members by descending alignment (SpscRing first for
  its cache-line alignment; the jthread stays last for RAII join-before-teardown)
  -> drops optin.performance.Padding with no behaviour change.
- GaplessSource / PureModePolicy / CrossfeedModule: rewrite raw loops as
  std::ranges::find_if / std::ranges::max / std::array::fill -> lets the cppcheck
  useStlAlgorithm blanket suppress be removed.
- strict-gate.sh: drop --suppress=useStlAlgorithm and the dead
  --suppress=missingIncludeSystem.

Deliberately kept (documented, not bare):
- DeviceBridge CDeviceInfo.name: 4-rule NOLINT. This file must be valid ISO
  C11, where std::array is not usable and every way to NAME the 256 bound trips
  a worse check (macro-to-enum / performance-enum-size wanting a C23 enum base
  type). Reason text corrected to match.
- .clang-tidy modernize-use-std-print: stays disabled with an updated rationale
  (production logs via the fwrite seam; std::print is gated on the Apple libc++
  deployment target, so enabling would push future logging to an unavailable API).

Verified: swift build; C++ null-test golden-master 0xe7267654ba01d315 UNCHANGED;
cppcheck (no --suppress) clean; clang-tidy clean across all 23 AudioDSP + test
TUs; clang-format clean; suppression-policy gate green.

+++ Sources/AudioDSP/include/AudioUnitBridge.h
@@ -11,12 +11,11 @@ namespace AdaptiveSound
     class AudioEngine;
     class DSPKernel;
 
-    /// C++-exported AU parameter IDs.
-    /// These correspond to AUAudioUnit parameter indexing.
-    /// Base type is uint64_t to match the C-ABI paramID width in setAUParameter/
-    /// getAUParameter (AudioUnitBridge.h); do not shrink it.
-    // NOLINTNEXTLINE(performance-enum-size) PERMANENT reason="width fixed to uint64_t to match the C-ABI paramID in set/getAUParameter"
-    enum class AUParameterID : uint64_t
+    /// C++-exported AU parameter IDs. The C-ABI paramID in set/getAUParameter is a SEPARATE raw
+    /// uint64_t; this enum is only used symbolically (never stored or cast as the ABI width), so it
+    /// takes the minimal underlying type. (Stage-2 suppression audit: the former uint64_t +
+    /// performance-enum-size NOLINT was unjustified — the enum is unused as a value.)
+    enum class AUParameterID : uint8_t
     {
         MasterGain = 0, // EQ master gain (dB)
         Bypass = 1,     // Bypass flag (0 = process, 1 = bypass)
